+++ app/database/migrations/2024_10_26_193600_create_servico_participacao_participantes_table.php
@@ -31,7 +31,7 @@ return new class extends Migration
             $table->uuid('participacao_tipo_id');
             $table->foreign('participacao_tipo_id', "{fk_{$this->model->getTableAsName()}_participacao_tipo_id")->references('id')->on((new App\Models\Tenant\ServicoParticipacaoTipoTenant)->getTableName());
 
-            $table->unsignedBigInteger('participacao_registro_tipo_id');
+            $table->smallInteger('participacao_registro_tipo_id');
             $table->foreign('participacao_registro_tipo_id', "{fk_{$this->model->getTableAsName()}_participacao_registro_tipo_id")->references('id')->on((new App\Models\Referencias\ParticipacaoRegistroTipo)->getTableName());
 
             $table->nullableUuidMorphs('referencia');
